requirements.txt
Flask==3.0.3
Flask_Cors==3.0.10
Flask_Limiter==3.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
supabase==2.7.4